+++ src/main/java/fr/eni/tp/spring_encheres/dal/RetraitDAO.java
@@ -3,7 +3,7 @@ package fr.eni.tp.spring_encheres.dal;
 import fr.eni.tp.spring_encheres.bo.Retrait;
 
 public interface RetraitDAO {
-    Retrait read(Integer idRetrait);
+    Retrait read(long idRetrait);
 
     void save(Retrait retrait);
 }
